+++ docs/src/App.fs
@@ -64,6 +64,7 @@ let menu currentPage =
               menuItem "Media" (Component Components.Media) currentPage
               menuItem "Menu" (Component Components.Menu) currentPage
               menuItem "Navbar" (Component Components.Navbar) currentPage
+              menuItem "Pagination" (Component Components.Pagination) currentPage
               menuItem "Message" (Component Components.Message) currentPage ] ]
 
 let header =
@@ -100,6 +101,7 @@ let root model dispatch =
             | Menu -> Components.Menu.View.root model.Components.Menu (MenuMsg >> dispatch)
             | Message -> Components.Message.View.root model.Components.Message (MessageMsg >> dispatch)
             | Navbar -> Components.Navbar.View.root model.Components.Navbar (NavbarMsg >> dispatch)
+            | Pagination -> Components.Pagination.View.root model.Components.Pagination (PaginationMsg >> dispatch)
 
     div []
         [ div [ ClassName "navbar-bg" ]